requirements.txt
pandas
numpy
pymongo
python-dotenv
requests
joblib
tensorflow
streamlit